feature-extractor/requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
numpy==1.19.5
opencv-python==4.7.0.72
Pillow==8.4.0
protobuf==3.19.6
six==1.16.0
